# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 汽车经销商 CRM API
  version: 1.0.0
  description: 当前后端 Controller 暴露接口契约。统一规则见 docs/api/api-conventions.md。
servers:
  - url: http://localhost:8089
    description: 本地后端服务
security:
  - BearerAuth: []
...
paths:
...
  /api/clues:
    get:
      tags: [Clue]
      operationId: listClues
      summary: 分页查询线索
      parameters:
        - $ref: "#/components/parameters/Current"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          $ref: "#/components/responses/PageOk"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/importExcel:
    post:
      tags: [Clue]
      operationId: importCluesFromExcel
      summary: 导入线索 Excel
      description: 导入逐行校验并归一化手机号；同文件或数据库中归一化后重复的手机号返回行级错误摘要。存在失败行时返回 422，但合法行可以通过 importedCount 表示已成功导入。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 仅支持 .xlsx，默认最大 5MB。
      responses:
        "200":
          description: 导入成功。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/R"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ImportResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          description: 文件可解析但存在行级导入错误，data 返回 ImportResult；合法行可能已经导入成功。
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/R"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ImportResult"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/clue:
    post:
      tags: [Clue]
      operationId: createClue
      summary: 新增线索
      description: 手机号先去除空格、横杠和括号等常见分隔符，再由生产和测试 Schema 的唯一约束兜底；并发重复手机号返回 409 DUPLICATE。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ClueQuery"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/ClueQuery"
      responses:
        "200":
          $ref: "#/components/responses/EmptyOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/ServerError"
    put:
      tags: [Clue]
      operationId: updateClue
      summary: 更新线索
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ClueQuery"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/ClueQuery"
      responses:
        "200":
          $ref: "#/components/responses/EmptyOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/clue/{phone}:
    get:
      tags: [Clue]
      operationId: checkCluePhone
      summary: 校验线索手机号是否可用
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: string
          description: 手机号；服务端按常见分隔符归一化后检查是否可用。
      responses:
        "200":
          $ref: "#/components/responses/EmptyOk"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [Clue]
      operationId: deleteClue
      summary: 删除线索
      description: 删除前校验客户引用；已转客户或存在客户引用的线索返回 422 RESOURCE_IN_USE，不删除线索备注和线索主体。
      parameters:
        - name: phone
          in: path
          required: true
          schema:
            type: integer
          description: 当前实现路径变量名为 phone，删除语义实际传线索 ID。
      responses:
        "200":
          $ref: "#/components/responses/EmptyOk"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/clue/detail/{id}:
    get:
      tags: [Clue]
      operationId: getClue
      summary: 查询线索详情
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          $ref: "#/components/responses/ClueObject"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/clue/{id}/owner:
    put:
      tags: [Clue]
      operationId: transferClueOwner
      summary: 转派线索负责人
      description: 转派必须提交目标负责人和原因；后端校验数据范围、目标负责人资格并写入责任历史。
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferClueOwnerRequest"
      responses:
        "200":
          $ref: "#/components/responses/EmptyOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/clue/{id}/owner-history:
    get:
      tags: [Clue]
      operationId: listClueOwnerHistory
      summary: 查询线索责任历史
      description: 查询前按当前用户数据范围校验线索访问权限。
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          $ref: "#/components/responses/ClueOwnerHistoryArrayOk"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/clue/{id}/close:
    put:
      tags: [Clue]
      operationId: closeClue
      summary: 关闭线索
      description: 关闭必须提交原因；后端校验数据范围、当前状态和影响行数，并写入操作审计。
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClueLifecycleRequest"
      responses:
        "200":
          $ref: "#/components/responses/EmptyOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/clue/{id}/restore:
    put:
      tags: [Clue]
      operationId: restoreClue
      summary: 恢复线索
      description: 恢复必须提交原因；后端校验数据范围、关闭状态、重复活跃线索和影响行数，并写入操作审计。
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClueLifecycleRequest"
      responses:
        "200":
          $ref: "#/components/responses/EmptyOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/clue/batch:
    post:
      tags: [Clue]
      operationId: batchDeleteClues
      summary: 批量删除线索
      description: 批量删除先完成全部访问权限和客户引用检查；任一线索被客户引用时整个批次返回 422 RESOURCE_IN_USE，不删除任何历史。
      requestBody:
        $ref: "#/components/requestBodies/IntegerArray"
      responses:
        "200":
          $ref: "#/components/responses/EmptyOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    Current:
      name: current
      in: query
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
      description: 历史分页页码。
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        default: 10
        minimum: 1
      description: 历史分页大小。
...
  requestBodies:
    IntegerArray:
      required: true
      content:
        application/json:
          schema:
            type: array
            items:
              type: integer
...
  responses:
    EmptyOk:
      description: 操作成功，无业务数据。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/R"
...
    ClueObject:
      description: 查询成功，data 为线索详情。
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/R"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Clue"
...
    ClueOwnerHistoryArrayOk:
      description: 查询成功，data 为线索责任历史数组。
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/R"
              - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ClueOwnerHistory"
...
    PageOk:
      description: 分页查询成功，data 为 PageInfo。
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/R"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/PageInfo"
...
    BadRequest:
      description: 请求参数错误。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorR"
    Unauthorized:
      description: 未认证、Token 无效或 Token 过期。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorR"
    Forbidden:
      description: 功能权限或数据权限不足。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorR"
    NotFound:
      description: 资源不存在。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorR"
    Conflict:
      description: 重复、并发冲突或非法状态迁移。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorR"
    Unprocessable:
      description: 请求格式正确但业务对象无法处理。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorR"
    ServerError:
      description: 未预期系统异常。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorR"
  schemas:
    R:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
          example: 操作成功
        data:
          nullable: true
    ErrorR:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          example: 501
        msg:
          type: string
          example: 请求参数格式有误
        data:
          nullable: true
    PageInfo:
      type: object
      properties:
        pageNum:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
        pages:
          type: integer
        list:
          type: array
          items:
            type: object
            additionalProperties: true
...
    ClueQuery:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        activityId:
          type: integer
        fullName:
          type: string
        appellation:
          type: integer
        phone:
          type: string
        weixin:
          type: string
        qq:
          type: string
        email:
          type: string
          format: email
        age:
          type: integer
        job:
          type: string
        yearIncome:
          type: number
        address:
          type: string
        needLoan:
          type: integer
        intentionState:
          type: integer
        intentionProduct:
          type: integer
        state:
          type: integer
        source:
          type: integer
        description:
          type: string
        nextContactTime:
          type: string
          format: date-time
    Clue:
      allOf:
        - $ref: "#/components/schemas/ClueQuery"
        - type: object
          properties:
            ownerDO:
              type: object
              additionalProperties: true
            activityDO:
              type: object
              additionalProperties: true
            appellationDO:
              type: object
              additionalProperties: true
            needLoanDO:
              type: object
              additionalProperties: true
            intentionStateDO:
              type: object
              additionalProperties: true
            intentionProductDO:
              type: object
              additionalProperties: true
            stateDO:
              $ref: "#/components/schemas/DictionaryValueSummary"
            sourceDO:
              type: object
              additionalProperties: true
    TransferClueOwnerRequest:
      type: object
      required: [newOwnerId, reason]
      properties:
        newOwnerId:
          type: integer
          description: 目标负责人用户 ID。
        reason:
          type: string
          description: 转派原因。
    ClueLifecycleRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          description: 关闭或恢复原因。
    ClueOwnerHistory:
      type: object
      properties:
        id:
          type: integer
        clueId:
          type: integer
        fromOwnerId:
          type: integer
        toOwnerId:
          type: integer
        assignedBy:
          type: integer
        reason:
          type: string
        assignedTime:
          type: string
          format: date-time
        fromOwnerName:
          type: string
        toOwnerName:
          type: string
        assignedByName:
          type: string
...
    DictionaryValueSummary:
      type: object
      properties:
        id:
          type: integer
        typeValue:
          type: string
        valueCode:
          type: string
...
    ImportResult:
      type: object
      properties:
        totalRows:
          type: integer
        validRows:
          type: integer
        failedRows:
          type: integer
        importedCount:
          type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              row:
                type: integer
              column:
                type: string
              reason:
                type: string
...